relocated and renamed the hclgen file to match existing file structure
